Get CLI serve command to work
This just leverages the vite dev server to be very similar to the `dev`
npm script. I used `yarn link` to test it with another local package and
found that I needed to use some absolute paths to ensure that the right
config got used. Next will be a similar treatment for the export
command.